# src/pkg/server/handler/apihandler/embedded/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventory Manager",
    "description": "Inventory Manager API",
    "version": "1.0.0",
    "contact": {
      "name": "LinuxCode",
      "email": "jbensch@linuxcode.net"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "inventory.linuxcode.net",
      "description": "Inventory server for example"
    }
  ],
...
  "paths": {
...
    "/api/inventories": {
      "get": {
        "summary": "Get all inventories",
        "description": "Retrieve all inventories",
        "operationId": "getAllInventories",
        "tags": [
          "inventory"
        ],
        "responses": {
          "200": {
            "description": "successful operation - returns a list of inventories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryListResponse"
                },
                "example": {
                  "inventories": [
                    {
                      "id": 1,
                      "name": "Backpack",
                      "userId": 1,
                      "max_weight": 850,
                      "volume": {
                        "height": 3,
                        "width": 4
                      }
                    },
                    {
                      "id": 2,
                      "name": "Chest",
                      "userId": 1,
                      "max_weight": 1500,
                      "volume": {
                        "height": 10,
                        "width": 10
                      }
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "pages": 10,
                    "limit": 10,
                    "total": 100,
                    "offset": 0
                  }
                }
              }
            }
          },
          "404": {
            "description": "No inventories found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": []
      },
      "post": {
        "summary": "Add new inventory",
        "description": "Add new inventory to the system",
        "operationId": "addInventory",
        "tags": [
          "inventory"
        ],
        "requestBody": {
          "required": true,
          "description": "Inventory to add to the system",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryPostRequest"
              },
              "example": {
                "name": "Backpack",
                "user_id": 1,
                "max_weight": 850,
                "volume": {
                  "height": 10,
                  "width": 10
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Added inventory - returns the newly created inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                },
                "example": {
                  "inventoryMeta": {
                    "id": 1,
                    "name": "Backpack",
                    "userId": 1,
                    "max_weight": 850,
                    "volume": {
                      "height": 3,
                      "width": 4
                    }
                  },
                  "items": [
                    {
                      "item": {
                        "id": 1,
                        "variant": "rare",
                        "name": "Sword",
                        "text": "A sword that can be used to fight tiny enemies.",
                        "buy_value": 100,
                        "sell_value": 50,
                        "weight": 10,
                        "durability": 100,
                        "max_stack": 1,
                        "shape": {
                          "rawshape": "###",
                          "height": 3,
                          "width": 1
                        },
                        "type": "meleeWeapon"
                      },
                      "position": {
                        "x": 0,
                        "y": 0,
                        "rotation": 0
                      },
                      "quantity": 1,
                      "durability_left": 70
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - e.g. invalid inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/inventories/{inventoryId}": {
      "get": {
        "summary": "Get inventory by ID",
        "description": "Retrieve an inventory by ID",
        "operationId": "getInventoryById",
        "tags": [
          "inventory"
        ],
        "parameters": [
          {
            "name": "inventoryId",
            "in": "path",
            "description": "ID of inventory to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation - returns inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                },
                "example": {
                  "inventoryMeta": {
                    "id": 1,
                    "name": "Backpack",
                    "userId": 1,
                    "max_weight": 850,
                    "volume": {
                      "height": 3,
                      "width": 4
                    }
                  },
                  "items": [
                    {
                      "item": {
                        "id": 1,
                        "variant": "rare",
                        "name": "Sword",
                        "text": "A sword that can be used to fight tiny enemies.",
                        "buy_value": 100,
                        "sell_value": 50,
                        "weight": 10,
                        "durability": 100,
                        "max_stack": 1,
                        "shape": {
                          "rawshape": "###",
                          "height": 3,
                          "width": 1
                        },
                        "type": "meleeWeapon"
                      },
                      "position": {
                        "x": 0,
                        "y": 0,
                        "rotation": 0
                      },
                      "quantity": 1,
                      "durability_left": 70
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied (e.g. not an integer) - can not find inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find inventory (inventory ID) to return"
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update an inventory",
        "description": "Update an inventory by ID",
        "operationId": "updateInventoryById",
        "tags": [
          "inventory"
        ],
        "parameters": [
          {
            "name": "inventoryId",
            "in": "path",
            "description": "ID of inventory to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Inventory object that needs to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryPostRequest"
              },
              "example": {
                "name": "Backpack",
                "user_id": 1,
                "max_weight": 850,
                "volume": {
                  "height": 10,
                  "width": 10
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update done - returns updated inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                },
                "example": {
                  "inventoryMeta": {
                    "id": 1,
                    "name": "Backpack",
                    "userId": 1,
                    "max_weight": 850,
                    "volume": {
                      "height": 3,
                      "width": 4
                    }
                  },
                  "items": [
                    {
                      "item": {
                        "id": 1,
                        "variant": "rare",
                        "name": "Sword",
                        "text": "A sword that can be used to fight tiny enemies.",
                        "buy_value": 100,
                        "sell_value": 50,
                        "weight": 10,
                        "durability": 100,
                        "max_stack": 1,
                        "shape": {
                          "rawshape": "###",
                          "height": 3,
                          "width": 1
                        },
                        "type": "meleeWeapon"
                      },
                      "position": {
                        "x": 0,
                        "y": 0,
                        "rotation": 0
                      },
                      "quantity": 1,
                      "durability_left": 70
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied (e.g. not an integer) - can not update inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find inventory (inventory ID) to update"
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete an inventory",
        "description": "Delete an inventory by ID",
        "operationId": "deleteInventoryById",
        "tags": [
          "inventory"
        ],
        "parameters": [
          {
            "name": "inventoryId",
            "in": "path",
            "description": "ID of inventory to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted inventory (inventory ID)"
          },
          "400": {
            "description": "Invalid ID supplied (e.g. not an integer) - can not delete inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find inventory (inventory ID) to delete"
          }
        },
        "security": []
      }
    },
    "/api/inventories/{inventoryId}/move": {
      "post": {
        "summary": "Move item in Inventory",
        "description": "Move an item already in the inventory",
        "operationId": "moveItemInInventory",
        "tags": [
          "inventory"
        ],
        "parameters": [
          {
            "name": "inventoryId",
            "in": "path",
            "description": "ID of inventory to move item in",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Position data and quantity to move",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveItemRequest"
              },
              "example": {
                "original_position": {
                  "x": 0,
                  "y": 0,
                  "rotation": 0
                },
                "new_position": {
                  "x": 1,
                  "y": 0,
                  "rotation": 0
                },
                "quantity": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Moved item in inventory"
          },
          "400": {
            "description": "Bad request (e.g. invalid item)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Can not find inventory or item in inventory"
          }
        },
        "security": []
      }
    },
    "/api/inventories/{inventoryId}/add": {
      "post": {
        "summary": "Add item to inventory (with optional position)",
        "description": "Adds a new item to the inventory with an optional position",
        "operationId": "addItemInInventory",
        "tags": [
          "inventory"
        ],
        "parameters": [
          {
            "name": "inventoryId",
            "in": "path",
            "description": "ID of inventory to add item to",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Item to add with quantity and optional position",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemToInventoryRequest"
              },
              "example": {
                "item": {
                  "id": 1,
                  "variant": "rare",
                  "name": "Sword",
                  "text": "Example of a rare sword.",
                  "buy_value": 100,
                  "sell_value": 50,
                  "weight": 10,
                  "durability": 100,
                  "max_stack": 1,
                  "shape": {
                    "rawshape": "###",
                    "height": 3,
                    "width": 1
                  },
                  "type": "meleeWeapon"
                },
                "quantity": 1,
                "durability_left": 100,
                "position": {
                  "x": 0,
                  "y": 0,
                  "rotation": 0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Added item to inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryItem"
                },
                "example": {
                  "item": {
                    "id": 1,
                    "variant": "rare",
                    "name": "Sword",
                    "text": "A sword that can be used to fight some enemies.",
                    "buy_value": 100,
                    "sell_value": 50,
                    "weight": 10,
                    "durability": 100,
                    "max_stack": 1,
                    "shape": {
                      "rawshape": "#####..#####",
                      "height": 3,
                      "width": 4
                    },
                    "type": "meleeWeapon"
                  },
                  "position": {
                    "x": 0,
                    "y": 0,
                    "rotation": 2
                  },
                  "quantity": 1,
                  "durability_left": 70
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g. invalid item or invalid position)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errorId": "123e4567-e89b-12d3-a456-426614174000",
                  "error": "The item does not exist",
                  "ErrorType": "Not Found",
                  "code": 404,
                  "timestamp": "2020-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "description": "This object holds the pagination data.",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "example": 100
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "example": 0
          }
        },
        "example": {
          "page": 1,
          "pages": 10,
          "limit": 10,
          "total": 100,
          "offset": 0
        },
        "required": [
          "page",
          "pages",
          "limit",
          "total",
          "offset"
        ]
      },
...
      "ErrorType": {
        "type": "string",
        "description": "The error type",
        "enum": [
          "Bad Request",
          "Not Found",
          "Not Implemented",
          "Internal Server Error"
        ],
        "example": "Bad Request"
      },
      "Error": {
        "type": "object",
        "description": "This object holds the error response data.",
        "properties": {
          "errorId": {
            "type": "string",
            "description": "The unique identifier for the error",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "error": {
            "type": "string",
            "description": "The error message",
            "example": "The item does not exist"
          },
          "ErrorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "code": {
            "type": "integer",
            "description": "The error code",
            "example": 404
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "The time the error occurred",
            "example": "2020-01-01T00:00:00Z"
          }
        },
        "example": {
          "errorId": "123e4567-e89b-12d3-a456-426614174000",
          "error": "The item does not exist",
          "ErrorType": "Not Found",
          "code": 404,
          "timestamp": "2020-01-01T00:00:00Z"
        },
        "required": [
          "errorId",
          "error",
          "ErrorType",
          "code",
          "timestamp"
        ]
      },
      "Volume": {
        "type": "object",
        "description": "This object holds the volume data.",
        "properties": {
          "height": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "width": {
            "type": "integer",
            "format": "int64",
            "example": 10
          }
        },
        "example": {
          "height": 10,
          "width": 10
        },
        "required": [
          "height",
          "width"
        ]
      },
...
      "InventoryMetaList": {
        "type": "array",
        "description": "The list of inventory metas.",
        "items": {
          "$ref": "#/components/schemas/InventoryMeta"
        },
        "example": [
          {
            "id": 1,
            "name": "Backpack",
            "userId": 1,
            "max_weight": 850,
            "volume": {
              "height": 3,
              "width": 4
            }
          },
          {
            "id": 2,
            "name": "Chest",
            "userId": 1,
            "max_weight": 1500,
            "volume": {
              "height": 10,
              "width": 10
            }
          }
        ]
      },
...
      "InventoryListResponse": {
        "type": "object",
        "description": "This object holds the inventory list response data. It includes the list of inventories and pagination data.",
        "properties": {
          "inventories": {
            "$ref": "#/components/schemas/InventoryMetaList"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "example": {
          "inventories": [
            {
              "id": 1,
              "name": "Backpack",
              "userId": 1,
              "max_weight": 850,
              "volume": {
                "height": 3,
                "width": 4
              }
            },
            {
              "id": 2,
              "name": "Chest",
              "userId": 1,
              "max_weight": 1500,
              "volume": {
                "height": 10,
                "width": 10
              }
            }
          ],
          "pagination": {
            "page": 1,
            "pages": 10,
            "limit": 10,
            "total": 100,
            "offset": 0
          }
        },
        "required": [
          "inventories",
          "pagination"
        ]
      },
      "InventoryPostRequest": {
        "type": "object",
        "description": "This object holds the request data for creating an inventory.",
        "properties": {
          "name": {
            "type": "string",
            "example": "Backpack"
          },
          "user_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "max_weight": {
            "type": "integer",
            "format": "int64",
            "example": 850
          },
          "volume": {
            "$ref": "#/components/schemas/Volume"
          }
        },
        "example": {
          "name": "Backpack",
          "user_id": 1,
          "max_weight": 850,
          "volume": {
            "height": 10,
            "width": 10
          }
        },
        "required": [
          "name",
          "user_id",
          "max_weight",
          "volume"
        ]
      },
...
      "InventoryItem": {
        "type": "object",
        "description": "This object holds the inventory item data.",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/Item"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          },
          "quantity": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "durability_left": {
            "type": "integer",
            "format": "int64",
            "example": 100
          }
        },
        "example": {
          "item": {
            "id": 1,
            "variant": "rare",
            "name": "Sword",
            "text": "A sword that can be used to fight some enemies.",
            "buy_value": 100,
            "sell_value": 50,
            "weight": 10,
            "durability": 100,
            "max_stack": 1,
            "shape": {
              "rawshape": "#####..#####",
              "height": 3,
              "width": 4
            },
            "type": "meleeWeapon"
          },
          "position": {
            "x": 0,
            "y": 0,
            "rotation": 2
          },
          "quantity": 1,
          "durability_left": 70
        },
        "required": [
          "item",
          "position",
          "quantity",
          "durability_left"
        ]
      },
...
      "Inventory": {
        "type": "object",
        "description": "This object holds the inventory meta data and the items in the inventory as an array of InventoryItem objects.",
        "properties": {
          "inventoryMeta": {
            "$ref": "#/components/schemas/InventoryMeta"
          },
          "items": {
            "$ref": "#/components/schemas/InventoryItemList"
          }
        },
        "example": {
          "inventoryMeta": {
            "id": 1,
            "name": "Backpack",
            "userId": 1,
            "max_weight": 850,
            "volume": {
              "height": 3,
              "width": 4
            }
          },
          "items": [
            {
              "item": {
                "id": 1,
                "variant": "rare",
                "name": "Sword",
                "text": "A sword that can be used to fight tiny enemies.",
                "buy_value": 100,
                "sell_value": 50,
                "weight": 10,
                "durability": 100,
                "max_stack": 1,
                "shape": {
                  "rawshape": "###",
                  "height": 3,
                  "width": 1
                },
                "type": "meleeWeapon"
              },
              "position": {
                "x": 0,
                "y": 0,
                "rotation": 0
              },
              "quantity": 1,
              "durability_left": 70
            }
          ]
        },
        "required": [
          "inventoryMeta",
          "items"
        ]
      },
      "MoveItemRequest": {
        "type": "object",
        "description": "This object holds the request data for moving an item within the inventory.",
        "properties": {
          "original_position": {
            "$ref": "#/components/schemas/Position"
          },
          "new_position": {
            "$ref": "#/components/schemas/Position"
          },
          "quantity": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        },
        "example": {
          "original_position": {
            "x": 0,
            "y": 0,
            "rotation": 0
          },
          "new_position": {
            "x": 1,
            "y": 0,
            "rotation": 0
          },
          "quantity": 1
        },
        "required": [
          "original_position",
          "new_position",
          "quantity"
        ]
      },
      "AddItemToInventoryRequest": {
        "type": "object",
        "description": "This object holds the request data for adding an item to the inventory.",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/Item"
          },
          "quantity": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "durability_left": {
            "type": "integer",
            "format": "int64",
            "example": 100
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          }
        },
        "example": {
          "item": {
            "id": 1,
            "variant": "rare",
            "name": "Sword",
            "text": "Example of a rare sword.",
            "buy_value": 100,
            "sell_value": 50,
            "weight": 10,
            "durability": 100,
            "max_stack": 1,
            "shape": {
              "rawshape": "###",
              "height": 3,
              "width": 1
            },
            "type": "meleeWeapon"
          },
          "quantity": 1,
          "durability_left": 100,
          "position": {
            "x": 0,
            "y": 0,
            "rotation": 0
          }
        },
        "required": [
          "item",
          "quantity",
          "durability_left"
        ]
      },
...
    }
  }
}